TD-775: Remove numpy dependency from sql.py
