+++ tests/How-To-Run-Test-And-How-To-Add-New-Test-Case.md
@@ -11,7 +11,7 @@
 4.  pip install ../src/connector/python/linux/python2 ; pip3 install
     ../src/connector/python/linux/python3
 
-5.  pip install numpy; pip3 install numpy    
+5.  pip install numpy; pip3 install numpy (numpy is required only if you need to run querySort.py)  
 
 >   Note: Both Python2 and Python3 are currently supported by the Python test
 >   framework. Since Python2 is no longer officially supported by Python Software
