+++ tests/pytest/util/sql.py
@@ -17,7 +17,6 @@ import time
 import datetime
 import inspect
 from util.log import *
-import numpy as np
 
 
 class TDSql:
@@ -199,47 +198,7 @@ class TDSql:
                 "%s failed: sql:%s, affectedRows:%d != expect:%d" %
                 (callerFilename, self.sql, self.affectedRows, expectAffectedRows))
         tdLog.info("sql:%s, affectedRows:%d == expect:%d" %
-                   (self.sql, self.affectedRows, expectAffectedRows))
-
-    def checkColumnSorted(self, col, order):
-        frame = inspect.stack()[1]
-        callerModule = inspect.getmodule(frame[0])
-        callerFilename = callerModule.__file__
-
-        if col < 0:
-            tdLog.exit(
-                "%s failed: sql:%s, col:%d is smaller than zero" %
-                (callerFilename, self.sql, col))
-        if col > self.queryCols:
-            tdLog.exit(
-                "%s failed: sql:%s, col:%d is larger than queryCols:%d" %
-                (callerFilename, self.sql, col, self.queryCols))
-
-        matrix = np.array(self.queryResult)
-        list = matrix[:, 0]
-
-        if order == "" or order.upper() == "ASC":
-            if all(sorted(list) == list):
-                tdLog.info(
-                    "sql:%s, column :%d is sorted in accending order as expected" %
-                    (self.sql, col))
-            else:
-                tdLog.exit(
-                    "%s failed: sql:%s, col:%d is not sorted in accesnind order" %
-                    (callerFilename, self.sql, col))
-        elif order.upper() == "DESC":
-            if all(sorted(list, reverse=True) == list):
-                tdLog.info(
-                    "sql:%s, column :%d is sorted in decending order as expected" %
-                    (self.sql, col))
-            else:
-                tdLog.exit(
-                    "%s failed: sql:%s, col:%d is not sorted in decending order" %
-                    (callerFilename, self.sql, col))
-        else:
-            tdLog.exit(
-                "%s failed: sql:%s, the order provided for col:%d is not correct" %
-                (callerFilename, self.sql, col))
+                   (self.sql, self.affectedRows, expectAffectedRows))    
 
 
 tdSql = TDSql()
